custom-elements/ElementInternals-role.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body></body>

<script>
function conditionally_register_element(role) {
    if (!customElements.get(`${role}-element`)) {
        customElements.define(`${role}-element`, class extends HTMLElement {
            constructor() {
                super();
                this.internals_ = this.attachInternals();
                this.internals_.role = role;
                this.internals_.ariaLabel = `Test ${role} label`;
            }
        });
    }
}

function generate_role_element(role) {
    conditionally_register_element(role);
    const el = document.createElement(`${role}-element`);
    document.body.append(el);
    return el;
}

function generate_child_role_element(role, parentElement) {
    conditionally_register_element(role);
    const el = document.createElement(`${role}-element`);
    parentElement.append(el);
    return el;
}

const test_roles = [
    'alert',
    'alertdialog',
    'application',
    'article',
    'banner',
    'button',
    'cell',
    'checkbox',
    'combobox',
    'complementary',
    'contentinfo',
    'definition',
    'dialog',
    ['directory', 'list'], // is this expected?
    'document',
    'figure',
    'form',
    'grid',
    'group',
    'heading',
    ['img', 'image'],
    'feed',
    'link',
    'list',
    'listbox',
    'log',
    'main',
    'marquee',
    'math',
    'menu',
    'menubar',
    'meter',
    'navigation',
    'note',
    'progressbar',
    'radio',
    'radiogroup',
    'region',
    'scrollbar',
    'search',
    'searchbox',
    'separator',
    'slider',
    'spinbutton',
    'status',
    'switch',
    'table',
    'tablist',
    'tabpanel',
    'term',
    'textbox',
    'timer',
    'toolbar',
    'tooltip',
    'tree',
    'treegrid'
];

test_roles.map((testdata) => {
    let role;
    let roleName;
    if (Array.isArray(testdata)) {
        role = testdata[0];
        roleName = testdata[1];
    } else {
        role = testdata;
        roleName = testdata;
    }
    promise_test(async () => {
        const el = generate_role_element(role);
        const computed_role = await test_driver.get_computed_role(el);
        const computed_label = await test_driver.get_computed_label(el);
        assert_equals(computed_role, roleName, el);
        assert_equals(computed_label, `Test ${role} label`, el);
        el.remove();
    }, `Applies "${role}" via Element Internals`);
});

const child_test_roles = [
    ['table', 'row', 'columnheader'],
    ['grid', 'row', 'gridcell'],
    ['list', 'listitem'],
    ['menu', 'menuitem'],
    ['menu', 'menuitemcheckbox'],
    ['menu', 'menuitemradio'],
    ['listbox', 'option'],
    ['table', 'row'],
    ['table', 'rowgroup'],
    ['table', 'row', 'rowheader'],
    ['tablist', 'tab'],
    ['tree', 'treeitem'],
];

child_test_roles.map((roles) => {
    promise_test(async () => {
        // Build the DOM tree by appending each role applied element as a child of the last.
        const el = generate_role_element(roles[0]);
        let currentElement = el;
        for (let i = 1; i < roles.length; i++) {
            currentElement = generate_child_role_element(roles[i], currentElement);
        }
        // Test the DOM tree starting at the initial ancestor of the tree.
        let testEl = el;
        for (let i = 0; i < roles.length; i++) {
            const role = roles[i];
            const computed_role = await test_driver.get_computed_role(testEl);
            const computed_label = await test_driver.get_computed_label(testEl);
            assert_equals(computed_role, role, testEl);
            assert_equals(computed_label, `Test ${role} label`, testEl);
            testEl = testEl.children[0];
        }
        el.remove();
    }, `Applies parent/child relationship of "${roles.join('"/"')}" via Element Internals`);
});
</script>
